refactor helper function to composable
