+++ demo-frontend/src/App.vue
@@ -6,7 +6,7 @@ import HelloWorld from "@/components/HelloWorld.vue";
 <template>
   <header class="container mx-auto flex items-center justify-between py-5 px-3">
     <div class="text-3xl">
-      <span class="text-gray-400">{</span>
+      <span class="text-gray-400">{ </span>
       <span class="text-red-600 text-4xl font-bold">A</span>nnotation
       <span class="text-red-600 text-4xl font-bold">W</span>eb
       <span class="text-red-600 text-4xl font-bold">S</span>ervice
